LayoutTests/http/tests/security/xssAuditor/malformed-xss-protection-header-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

This tests that the X-XSS-Protection header is not ignored when the length of its value exceeds 16 characters.



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
